remove http error tests
